Show the localized error description for Swift errors
We were loosing user-friendly error strings with the `String(describing:)` approach and there was no way to work around this as an API user. This change now makes it possible for us to conform project-specific errors to `LocalizedError` and we can get richer errors.

This change includes a couple of drive-by warning fixes and one addition of `()` to quiet the compiler warning where we know better.