Remove LuaStateStringEncoding.swift, move default to registry
It is preferable to use the registry rather than lua_getextraspace to
store default string encoding preference, because it works with
lua_States from anywhere (which might be using the extraspace
themselves). And not using something nonportable means we can roll it
into LuaState.swift, which makes the API much neater.

For example, String is now always Pushable.